% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visdatid-package.R
\docType{package}
\name{visdatid-package}
\alias{visdatid}
\alias{visdatid-package}
\title{visdatid: Visual Data Quality Diagnostics and Exploration}
\description{
The \code{visdatid} package extends visual data exploration concepts from
the \code{visdat} package. It provides advanced tools for diagnosing missing
values, numerical distributions, correlations, and general data-quality
problems.

The package was developed as an attributed derivative work of the
\code{visdat} package for an educational R package development project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mona910/visdatid}
  \item \url{https://github.com/ropensci/visdat}
  \item Report bugs at \url{https://github.com/mona910/visdatid/issues}
}

}
\author{
\strong{Maintainer}: Syifa Nabila Motumona \email{24611064@students.ac.id}

Authors:
\itemize{
  \item Syifa Nabila Motumona \email{24611064@students.ac.id}
  \item Ermadina Mutiara \email{24611012@students.ac.id}
  \item M. Arif Al-Ghifary \email{24611049@students.ac.id}
  \item Ahmad Faiz Muzaki \email{24611023@students.ac.id}
}

}
\keyword{internal}
